edit hetero nn benchmark
