--- examples/pipeline/hetero_nn/hetero_nn_breast_config.yaml
@@ -1,17 +1,0 @@
-data_guest: "examples/data/breast_hetero_guest.csv"
-data_host: "examples/data/breast_hetero_host.csv"
-idx: "id"
-label_name: "y"
-loss: "binary_crossentropy"
-learning_rate: 0.15
-tol: 0.01
-batch_size: 320
-metrics: ["AUC"]
-epochs: 100
-bottom_layer_units: 2
-interactive_layer_units: 2
-top_layer_units: 1
-guest_table_name: "breast_hetero_guest"
-host_table_name: "breast_hetero_host"
-top_act: "sigmoid"
-opt: "SGD"
