--- examples/pipeline/hetero_nn/hetero_nn_default_credit_config.yaml
@@ -1,18 +1,0 @@
-data_guest: "examples/data/default_credit_hetero_guest.csv"
-data_host: "examples/data/default_credit_hetero_host.csv"
-idx: "id"
-label_name: "y"
-learning_rate: 0.05
-loss: "binary_crossentropy"
-tol: 0.01
-batch_size: 30000
-metrics: ["AUC"]
-epochs: 20
-bottom_layer_units: 2
-interactive_layer_units: 1
-top_layer_units: 1
-guest_table_name: "default_credit_hetero_guest"
-host_table_name: "default_credit_hetero_host"
-top_act: "sigmoid"
-opt: "RMSprop"
-
