--- examples/pipeline/hetero_nn/hetero_nn_epsilon_5k_config.yaml
@@ -1,17 +1,0 @@
-data_guest: "examples/data/epsilon_5k_hetero_guest.csv"
-data_host: "examples/data/epsilon_5k_hetero_host.csv"
-idx: "id"
-label_name: "y"
-learning_rate: 0.03
-loss: "binary_crossentropy"
-tol: 0.01
-batch_size: 2048
-metrics: ["AUC"]
-epochs: 100
-bottom_layer_units: 2
-interactive_layer_units: 1
-top_layer_units: 1
-guest_table_name: "epsilon_5k_hetero_guest"
-host_table_name: "epsilon_5k_hetero_host"
-top_act: "sigmoid"
-opt: "RMSprop"
