--- examples/pipeline/hetero_nn/hetero_nn_give_credit_config.yaml
@@ -1,17 +1,0 @@
-data_guest: "examples/data/give_credit_hetero_guest.csv"
-data_host: "examples/data/give_credit_hetero_host.csv"
-idx: "id"
-label_name: "y"
-learning_rate: 0.05
-tol: 0.01
-loss: "binary_crossentropy"
-batch_size: 15000
-metrics: ["AUC"]
-epochs: 5
-bottom_layer_units: 2
-interactive_layer_units: 1
-top_layer_units: 1
-guest_table_name: "give_credit_hetero_guest"
-host_table_name: "give_credit_hetero_host"
-top_act: "sigmoid"
-opt: "RMSprop"
